Singleton pattern in model and controller

+++ core/src/com/aor/arena/model/entities/BulletModel.java
@@ -32,7 +32,7 @@ public class BulletModel extends EntityModel{
      * Sets this bullet's time to live in seconds
      * @param timeToLive
      */
-    public void setTimeToLive(int timeToLive) {
+    public void setTimeToLive(float timeToLive) {
         this.timeToLive = timeToLive;
     }
 }
